+++ pkg/source/traderjoecommon/type.go
@@ -1,0 +1,5 @@
+package traderjoecommon
+
+type Metadata struct {
+	Offset int `json:"offset"`
+}
