--- pkg/source/traderjoev21/config.go
@@ -1,7 +1,0 @@
-package traderjoev21
-
-type Config struct {
-	DexID          string `json:"dexID"`
-	FactoryAddress string `json:"factoryAddress"`
-	NewPoolLimit   int    `json:"newPoolLimit"`
-}
