+++ pkg/source/traderjoev21/pool_tracker.go
@@ -2,53 +2,33 @@ package traderjoev21
 
 import (
 	"context"
-	"time"
 
 	"github.com/KyberNetwork/ethrpc"
 	"github.com/KyberNetwork/logger"
 
 	"github.com/KyberNetwork/kyberswap-dex-lib/pkg/entity"
+	"github.com/KyberNetwork/kyberswap-dex-lib/pkg/source/traderjoecommon"
 )
 
 type PoolTracker struct {
-	ethrpcClient *ethrpc.Client
+	*traderjoecommon.PoolTracker[Reserves]
 }
 
 func NewPoolTracker(
 	ethrpcClient *ethrpc.Client,
 ) (*PoolTracker, error) {
 	return &PoolTracker{
-		ethrpcClient: ethrpcClient,
+		PoolTracker: &traderjoecommon.PoolTracker[Reserves]{
+			EthrpcClient:          ethrpcClient,
+			PairABI:               pairABI,
+			PairGetReservesMethod: pairGetReservesMethod,
+		},
 	}, nil
 }
 
 func (d *PoolTracker) GetNewPoolState(ctx context.Context, p entity.Pool) (entity.Pool, error) {
-	logger.Infof("[TraderJoe v2.1] Start getting new state of pool: %v", p.Address)
-
-	rpcRequest := d.ethrpcClient.NewRequest()
-	rpcRequest.SetContext(ctx)
-
-	var reserves Reserves
-
-	rpcRequest.AddCall(&ethrpc.Call{
-		ABI:    pairABI,
-		Target: p.Address,
-		Method: pairGetReservesMethod,
-	}, []interface{}{&reserves})
-
-	_, err := rpcRequest.Call()
-	if err != nil {
-		logger.Errorf("failed to call pool: %v, err: %v", p.Address, err)
-		return entity.Pool{}, err
-	}
-
-	p.Timestamp = time.Now().Unix()
-	p.Reserves = entity.PoolReserves{
-		reserves.ReserveX.String(),
-		reserves.ReserveY.String(),
-	}
-
-	logger.Infof("[TraderJoe v2.1] Finish getting new state of pool: %v", p.Address)
-
-	return p, nil
+	logger.Infof("[TraderJoe v2.0] Start getting new state of pool: %v", p.Address)
+	p, err := d.PoolTracker.GetNewPoolState(ctx, p)
+	logger.Infof("[TraderJoe v2.0] Finish getting new state of pool: %v", p.Address)
+	return p, err
 }
