+++ pkg/source/traderjoev21/type.go
@@ -1,6 +1,10 @@
 package traderjoev21
 
-import "math/big"
+import (
+	"math/big"
+
+	"github.com/KyberNetwork/kyberswap-dex-lib/pkg/entity"
+)
 
 // Reserves https://github.com/traderjoe-xyz/joe-v2/blob/v2.1.1/src/LBPair.sol#L160
 type Reserves struct {
@@ -8,6 +12,9 @@ type Reserves struct {
 	ReserveY *big.Int
 }
 
-type Metadata struct {
-	Offset int `json:"offset"`
+func (r Reserves) GetPoolReserves() entity.PoolReserves {
+	return entity.PoolReserves{
+		r.ReserveX.String(),
+		r.ReserveY.String(),
+	}
 }
